requirements.txt
pymongo==4.10.1
apache-airflow-providers-mongo==5.0.0